+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ### Added
 - New configuration option `copyPythonPath.omitRootPath` to omit specific root paths from the dotted path
+- New functionality to copy the full dotted path of imported symbols
+- Support for generating import statements for imported symbols
 
 ## [0.6.9]
 - Initial release